requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
idna==2.6
requests==2.18.4
six==1.11.0
slack-sansio==0.3.2
urllib3==1.22
websocket-client==0.46.0
